# docs/swagger.yaml
definitions:
  models.Book:
    properties:
      author:
        description: Book author
        example: Mr. Cccc Dddd
        type: string
      category:
        description: Book category
        example: Sea
        type: string
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      id:
        description: Model id pk
        type: integer
      name:
        description: Book name
        example: Mr. Aaaa Bbbbb
        type: string
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
    type: object
  models.BookForCreate:
    properties:
      author:
        description: Book author
        example: Mr. Cccc Dddd
        type: string
      category:
        description: Book category
        example: Sea
        type: string
      name:
        description: Book name
        example: Mr. Aaaa Bbbbb
        type: string
    type: object
  models.BookForUpdate:
    properties:
      author:
        description: Book author
        example: Mr. Cccc Dddd
        type: string
      category:
        description: Book category
        example: Sea
        type: string
      name:
        description: Book name
        example: Mr. Aaaa Bbbbb
        type: string
    type: object
...
  utils.ResponseData:
    properties:
      status:
        description: Response status
        type: integer
      success:
        default: true
        description: Response success
        example: true
        type: boolean
    type: object
info:
  contact:
    email: limweb@hotmail.com
    name: API Support
    url: http://www.servit.co.th/support
  description: '# Manageing Customer API'
  license:
    name: The MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://www.servit.co.th/
  title: ApiTest
  version: "1.0"
paths:
...
  /api/v1/books/all:
    get:
      consumes:
      - application/json
      description: List all books
      operationId: ListBooks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Book'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: List Books
      tags:
      - books
  /api/v1/books/by/{id}:
    get:
      consumes:
      - application/json
      description: Get book by id
      operationId: GetBook
      parameters:
      - description: id of book to be gotten
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Basic auth: []
      summary: Get Book
      tags:
      - books
  /api/v1/books/del/{id}:
    get:
      consumes:
      - application/json
      description: Delete book by id
      operationId: DeleteBook
      parameters:
      - description: id of book to be deleted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Delete Book
      tags:
      - books
  /api/v1/books/edit/{id}:
    post:
      consumes:
      - application/json
      description: Update book by id
      operationId: UpdateBook
      parameters:
      - description: id of book to be updated
        in: path
        name: id
        required: true
        type: integer
      - description: Book data to be updated
        in: body
        name: Book
        required: true
        schema:
          $ref: '#/definitions/models.BookForUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Update Book
      tags:
      - books
  /api/v1/books/last:
    get:
      consumes:
      - application/json
      description: List all books
      operationId: ListlastBooks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Book'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: List Books
      tags:
      - books
  /api/v1/books/new:
    post:
      consumes:
      - application/json
      description: Create new book
      operationId: CreateBook
      parameters:
      - description: Book data to be created
        in: body
        name: Book
        required: true
        schema:
          $ref: '#/definitions/models.BookForCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Create Book
      tags:
      - books
  /api/v1/books/vuetable:
    get:
      consumes:
      - application/json
      description: List all books with pagination keyword filter by column Sort columns
      operationId: vuetableBook
      parameters:
      - description: number of page
        in: query
        name: page
        type: string
      - description: get number of per_page
        in: query
        name: per_page
        type: string
      - description: 'soft by cols exm: col1|asc,col2|desc'
        in: query
        name: sort
        type: string
      - description: 'search with column exm: col1|aaa,col2|bbb  by filter'
        in: query
        name: filter
        type: string
      - description: 'search by kw exm: aaa bbb ccc '
        in: query
        name: kw
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: List Books use VueTable
      tags:
      - books
...
schemes:
- https
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"